Moved Send methods to message building helpers.
