+++ src/login/messages/login_client_token.cc
@@ -1,0 +1,37 @@
+#include "login/messages/login_client_token.h"
+
+#include "login/login_client.h"
+#include "login/account.h"
+
+using namespace login;
+using namespace messages;
+using namespace std;
+
+using anh::ByteBuffer;
+
+LoginClientToken login::messages::BuildLoginClientToken(shared_ptr<LoginClient> login_client) {
+    static const uint8_t data[56] = {
+        0x20, 0x00, 0x00, 0x00, 
+        0x15, 0x00, 0x00, 0x00,
+        0x0E, 0xD6, 0x93, 0xDE, 
+        0xD2, 0xEF, 0xBF, 0x8E,
+        0xA1, 0xAC, 0xD2, 0xEE, 
+        0x4C, 0x55, 0xBE, 0x30,
+        0x5F, 0xBE, 0x23, 0x0D, 
+        0xB4, 0xAB, 0x58, 0xF9,
+        0x62, 0x69, 0x79, 0x67, 
+        0xE8, 0x10, 0x6E, 0xD3,
+        0x86, 0x9B, 0x3A, 0x4A, 
+        0x1A, 0x72, 0xA1, 0xFA,
+        0x8F, 0x96, 0xFF, 0x9F, 
+        0xA5, 0x62, 0x5A, 0x29,
+    };
+    
+    LoginClientToken message;
+
+    message.station_id = login_client->account->account_id();
+    message.session_key = ByteBuffer(data, sizeof(data));
+    message.station_username = login_client->username;
+
+    return message;
+}
